context/requirements.txt
fastapi
uvicorn
redis
python-dotenv
langchain
langchain-community
langchain-openai
langchain-redis